util/requirements.txt
pandas>=2.1.0
reportlab>=4.2.5
google-cloud-storage>=2.18.2